+++ src/lib.rs
@@ -14,10 +14,11 @@
 //! - Wrappers around assembly instructions like `WFI`.
 
 #![no_std]
+#![feature(asm_const)]
 
 pub mod asm;
-pub mod delay;
-pub mod interrupt;
+//pub mod delay;
+//pub mod interrupt;
 pub mod register;
 
 #[macro_use]